pyproject.toml
[project]
name = "aistrak"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.5",
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "certifi==2025.8.3",
    "clerk-backend-api>=3.3.0",
    "click==8.2.1",
    "dnspython==2.7.0",
    "email-validator==2.2.0",
    "fastapi==0.116.1",
    "fastapi-cli==0.0.8",
    "fastapi-cloud-cli==0.1.5",
    "greenlet==3.2.3",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "mysql-connector-python>=9.4.0",
    "mysqlclient==2.2.7",
    "orjson==3.11.1",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-extra-types==2.10.5",
    "pydantic-settings==2.10.1",
    "pygments==2.19.2",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "requests>=2.32.5",
    "rich==14.1.0",
    "rich-toolkit==0.14.9",
    "rignore==0.6.4",
    "sentry-sdk==2.34.1",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.42",
    "starlette==0.47.2",
    "svix>=1.75.0",
    "typer==0.16.0",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "ujson==5.10.0",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "uvloop==0.21.0",
    "watchfiles==1.1.0",
    "websockets==15.0.1",
]
